Wire up SPU emulator and sound system initialization
- Call spu_emu_init(), sd_mem_alloc(), sd_init() during game init
- Replace PSX hardcoded address 0x801E0000 with allocated buffer
- Set master volume after sd_init (which zeros it)
- Call IntSdMain 16x per frame to match PSX interrupt rate
- Add .wvx and .mdx loading in port FS stage loader
- Fix 64-bit pointer truncation in SD_80083F54
- Prevent double SPU init

Sound doesn't produce audio yet because PcmOpen/PcmRead/PcmClose
(disc file loading) are unimplemented stubs returning -1, so wave
data never reaches SPU RAM.

+++ source/sound/sd_file.c
@@ -416,7 +416,7 @@ int SD_80083F54(char *end)
 
     wave_load_ptr = cdload_buf + 16;
 
-    if (((unsigned int)src_ptr + size) >= (unsigned int)end)
+    if (((uintptr_t)src_ptr + size) >= (uintptr_t)end)
     {
         return 0;
     }
